Fix bundler related dev anti-pattern in UAT
  Our user acceptance test suite should not know about the program
source code or be able to reuse it. However, since we "facilitated"
bundler usage for travis CI, it is enabled during tests. bundler breaks
current `RUBYOPT' value and rely on it to inject code modifying the
runtime behavior of executed ruby programs. Some basic ruby core method
are changed, and many unwanted libraries are loaded.

  The consequences are that when aruba run `uhwm', it does not behave as
intended, and `uh/wm' will be loaded in the UAT suite, even if we didn't
require it.